Updated: Moved quick action helper funtion to component
